Adds Table of Content to... well, Table of Content
I don't know why, but it's kind of funny to me to see Table of Content listed in Table of Content, so I thought I should submit that as a pull request.

### Note to other contributors

Please let @sindresorhus decide whether or not to accept this. I'm submitting this as a pull request because I don't want to make him mad.