Add link in Learn: Dive into Machine Learning
Hey there. Love the Awesome lists and the Awesome List of Awesome Lists :)

I feel like there is a place for my guide here. Hope you feel the same.

Regarding which section to put it in ... I put it in "Learn" since it really is introductory (very tightly curated so as to NOT be overwhelming). The other lists in the "Machine Learning" section are more exhaustive. So I think this belongs in "Learn." Let me know!