Adding tips repository to rails category
Whenever I use to search for a best one and most preferred ones to use, I come down to awesome and check the list available and look into those. By that way I came to see the jstips, which was remarkably useful. So I thought of searching for tips related to rails and there were few but not maintained. So, I thought of making this for rails category too. And I discussed with few more devs and they too felt useful and so I am proposing this.