Add `Funny Markov`
This is a funny collection of Markov chain output: https://github.com/sublimino/awesome-funny-markov

I've attempted to comply with the manifesto and contribution guidelines (thanks for maintaining such a huge list).

I believe this is applicable to the `awesome-list` concept as funny Markov chains are an integral part of computer history, sitting at the rare intersection of technology and laughter. A [history](https://github.com/sublimino/awesome-funny-markov/blob/master/README.md#a-history-of-funny-markov-chains) section gives some background.