Add a link to Yeoman generator for awesome lists
I wrote this Yeoman generator to make it easy to get started building an awesome list, and thought maybe the readers of this doc might find it useful. The generated directory structure looks [like this](https://github.com/dar5hak/generator-awesome-list/tree/master/app/templates).

Is it okay to add it here?