Explorable explanations list
[Link to the list](https://github.com/sp4ke/awesome-explorables)

This is a curated list of explorable explanations resources. Explorable Explanations as defined by Bret Victor are rich interactive documents that allow the reader to explore and question the topic while learning. This list is officially supported by the /r/explorables reddit.

I am the maintainer of the list.  I made sure to read the contribution guidelines many times so please let me know if there are mistakes, last time I tried to submit this list it was just denied without further details.