Add awesome-monorepo.
Monorepos, short for monolithic repositories, large source-control repositories where multiple projects cohabitate, have been kind of a trend in the past 2 years or so. Everyone is talking about how everyone else is doing it. I've been collecting resources on them for the past 6 months or so, so I think it's time to add this list to the list of list.