Add Open Computational Neuroscience Resources
I would like to add my curated list of computational neuroscience resource tools. There is much fragmentation of methods and tools in the computational neuroscience community and this is an attempt to compile them. That said, this also means there are a great many ways anyone can contribute to the field or learn from it with nothing more than a computer! These are very accessible resources for anyone desiring to learn about the brain.