Add CircuitPython Awesome list to Python Group
Request to add the Adafruit Industries curated awesome list devoted to CircuitPython, a microcontroller based version of Python, to the Python category. The Adafruit CircuitPython awesome list is at https://github.com/adafruit/awesome-circuitpython

[https://github.com/adafruit/awesome-circuitpython]

[Request to add the Adafruit Industries curated awesome list devoted to CircuitPython, a microcontroller based version of Python, to the Python category.]

By submitting this pull request I confirm I've read and complied with the below requirements 
Please read it multiple times. I spent a lot of time on these guidelines and most people miss a lot.

Requirements for your pull request
•	I have read and understood the contribution guidelines and the instructions for creating a list.
•	This pull request has a descriptive title.
For example, Add Name of List, not Update readme.md or Add awesome list.
•	The entry in the Awesome list should:
o	Include a short description about the project/theme of the list. It should not describe the list itself.
Example: - [Fish](…) - User-friendly shell., not - [Fish](…) - Resources for Fish..
o	Be added at the bottom of the appropriate category.
•	The list I'm submitting complies with these requirements:
Requirements for your Awesome list
•	Has been around for at least 30 days.
That means 30 days from either the first real commit or when it was open-sourced. Whatever is most recent.
•	It's the result of hard work and the best I could possibly produce.
•	Non-generated Markdown file in a GitHub repo.
•	Includes a succinct description of the project/theme at the top of the readme. (Example)
•	The repo should have awesome-list & awesome as GitHub topics. I encourage you to add more relevant topics.
•	Not a duplicate.
•	Only has awesome items. Awesome lists are curations of the best, not everything.
•	Includes a project logo/illustration whenever possible.
o	Either fullwidth or placed at the top-right of the readme. (Example)
o	The image should link to the project website or any relevant website.
o	The image should be high-DPI. Set it to maximum half the width of the original image.
•	Entries have a description, unless the title is descriptive enough by itself. It rarely is though.
•	Includes the Awesome badge.
o	Should be placed on the right side of the readme heading.
o	Should link back to this list.
•	Has a Table of Contents section.
o	Should be named Contents, not Table of Contents.
o	Should be the first section in the list.
o	Should only have one level of sub-lists, preferably none.
•	Has an appropriate license.
o	That means something like CC0, not a code licence like MIT, BSD, Apache, etc.
o	WTFPL and Unlicense are not acceptable licenses.
o	If you use a license badge, it should be SVG, not PNG.
•	Has contribution guidelines.
o	The file should be named contributing.md. Casing is up to you.
•	Has consistent formatting and proper spelling/grammar.
o	The link and description are separated by a dash. 
Example: - [AVA](…) - JavaScript test runner.
o	The description starts with an uppercase character and ends with a period.
o	Consistent and correct naming. For example, Node.js, not NodeJS or node.js.
•	Doesn't include a Travis badge.
You can still use Travis for list linting, but the badge has no value in the readme.
Go to the top and read it again.