Suggest new awesome list: Cybersecurity Blue Team.
This list is focused on curating a big picture overview of tools and
resources dealing with collaborative security flaw identification,
remediation, and verification. It is a relatively new list that provides
a holistic subset of several other security-related lists along with a
number of new items focused primarily around the needs of DevSecOps/Blue
Team needs in an organization. Relevant other lists, such as awesome IR,
awesome pentest, etc., are linked where relevant so as not to duplicate
work or reinvent the wheel. (Instead, I contribute to those other
lists directly when more appropriate than adding to this list.)

Thanks for the consideration.