Add Artificial General Intelligence
https://github.com/fairy-tale-agi-solutions/awesome-artificial-general-intelligence

Artificial general intelligence (AGI) is the intelligence of a machine that could successfully perform any intellectual task that a human being can. It is a primary goal of some artificial intelligence research and a common topic in science fiction and future studies. Artificial general intelligence is also referred to as "strong AI", "full AI" or as the ability of a machine to perform "general intelligent action".

Pull requests I reviewed:

https://github.com/sindresorhus/awesome/pull/1407
https://github.com/sindresorhus/awesome/pull/1420
https://github.com/sindresorhus/awesome/pull/1415
https://github.com/sindresorhus/awesome/pull/1414