Add Fuzzing
https://github.com/cpuu/awesome-fuzzing/
Add Fuzzing for security testing

By submitting this pull request I confirm I've read and complied with the below requirements vulcan_salute
Please read it multiple times. I spent a lot of time on these guidelines and most people miss a lot.

Requirements for your pull request
You have to review at least 2 other open pull requests. Try to prioritize unreviewed PRs, but you can also add more comments to reviewed PRs. Go through the below list when reviewing. This requirement is meant to help make the Awesome project self-sustaining. Comment here which PRs you reviewed. You're expected to put a good effort into this and to be thorough. Look at previous PR reviews for inspiration.
I have read and understood the instructions for creating a list.
This pull request has a descriptive title in the format Add Name of List (Example: Add Swift), not Update readme.md or Add awesome list.
The entry in the Awesome list should:
Include a short description about the project/theme of the list. It should not describe the list itself.
Example: - [Fish](…) - User-friendly shell., not - [Fish](…) - Resources for Fish..
Be added at the bottom of the appropriate category.
The list I'm submitting complies with the below requirements.
Requirements for your Awesome list
Has been around for at least 30 days.
That means 30 days from either the first real commit or when it was open-sourced. Whatever is most recent.
It's the result of hard work and the best I could possibly produce.
Non-generated Markdown file in a GitHub repo.
Includes a succinct description of the project/theme at the top of the readme. (Example)
The repo should have awesome-list & awesome as GitHub topics. I encourage you to add more relevant topics.
Not a duplicate.
Only has awesome items. Awesome lists are curations of the best, not everything.
Includes a project logo/illustration whenever possible.
Either centered, fullwidth, or placed at the top-right of the readme. (Example)
The image should link to the project website or any relevant website.
The image should be high-DPI. Set it to maximum half the width of the original image.
Entries have a description, unless the title is descriptive enough by itself. It rarely is though.
Includes the Awesome badge.
Should be placed on the right side of the readme heading.
Can be placed centered if the list has a centered graphics header.
Should link back to this list.
Has a Table of Contents section.
Should be named Contents, not Table of Contents.
Should be the first section in the list.
Should only have one level of nested lists, preferably none.
Has an appropriate license.
That means something like CC0, not a code licence like MIT, BSD, Apache, etc.
WTFPL and Unlicense are not acceptable licenses.
If you use a license badge, it should be SVG, not PNG.
Has contribution guidelines.
The file should be named contributing.md. Casing is up to you.
Has consistent formatting and proper spelling/grammar.
The link and description are separated by a dash. 
Example: - [AVA](…) - JavaScript test runner.
The description starts with an uppercase character and ends with a period.
Consistent and correct naming. For example, Node.js, not NodeJS or node.js.
Doesn't include a Travis badge.
You can still use Travis for list linting, but the badge has no value in the readme.
Go to the top and read it again.