update readme.md
support for awesome-django was stopped by 'Roberto Rosario' due to some personal problems mentioned here "https://gitlab.com/rosarior/awesome-django" due to which the broken link was removed but being a django developer I considered it a great resource and wanted for other developers to still have access to it, so I have decided to take it further from here. Please consider this link and I will try my level best to keep it in a great shape.