Platforms / Azure Cloud / Use Case Cheat Sheet
While building a scalable system is lot more easier with the advent of cloud technologies, identifying the right fit can be challenging. Usually it may take years of solid experience in architectural space.

Added the link to the repo that contains a curated list of cloud services as available in azure to be used in real world use case scenarios.