Update readme.md
Hi there, I'm Daniel, I'm from Deepnote. We're a small startup creating a better data science notebook and one of our users made this super cool repo with a list of interesting projects in Deepnote, so I thought it would be nice to add it here :)
Btw thanks for keeping this list, it's an awesome source of great projects!