Change link from sysadmin repo to sysadmin forum
Hi, I decided to archive [my awesome-sysadmin list](https://github.com/n1trux/awesome-sysadmin) and move it to a Flarum installation I manage:

https://awesome.tilde.fun/t/sysadmin

While this certainly means that the list can be inaccessible at any time, I try to make sure to host it indefinitely. Flarum uses Markdown format and I'm sure we can find a way to mirror this to a git repository somehow.

The reason for my decision is that searching via tags and discussing the software is way easier in Flarum than it is in a Github markdown list. I will try to create an extension to mirror the layout/design of the initial markdown list with a powerful search interface, but for now I think Flarum's search features are good enough™.

I'll enable Github OAuth access so everyone with a Github account can log in into the awesome-sysadmin forum and contribute there.

Tagging @nodiscc and @Kickball for consideration.