Add Quantitative Finance
Quantitative finance is the use of mathematical models to analyze financial markets. This list aims to make it easier for students and professionals to find valuable information, tools and techniques related to the theory of mathematical finance. I hope this list will help people quickly find the best resources available for a specific subject.